backend 권한 및 task에 대한 수정, 삭제 권한수정 * admin은 모든 task에 대해서 변경 및 수정, 삭제가 가능하다. * admin을 제외한 모든 사용자는 자신의 task에 대해서만 수정,삭제 권한을 가진다. * 모든 로그인 사용자는 자신이 작성한 task 이외의 task도 조회 할수 있다.

+++ src/main/java/com/itplatform/todo/service/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package com.itplatform.todo.service;
 
 import com.itplatform.todo.domain.Task;
+import com.itplatform.todo.domain.User;
 import com.itplatform.todo.repository.TaskRepository;
 import java.util.List;
 import java.util.Optional;
@@ -53,17 +54,27 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     @Transactional
-    public Task update(Task task) {
-        if (!taskRepository.existsById(task.getId())) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not updated");
-        }
+    public Task update(Task task, User user) {
+        Task orgTask = taskRepository.findById(task.getId()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not updated"));
+
+        if (!orgTask.getReporter().equals(task.getReporter()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 접근 입니다.");
+
+        if (!(task.getReporter().equals(user.getName()) || user.isAdmin()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "수정 권한이 없습니다.");
 
         return taskRepository.save(task);
     }
 
     @Override
     @Transactional
-    public void deleteById(int id) {
+    public void deleteById(int id, User user) {
+
+        Task orgTask = taskRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found task"));
+        if (!(orgTask.getReporter().equals(user.getName()) || user.isAdmin()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "수정 권한이 없습니다.");
+
         taskRepository.deleteById(id);
     }
+
 }
